requirements.txt
# UI
streamlit>=1.32.0

# Agents & graph
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
openai>=1.40.0

# MCP
mcp>=1.0.0

# PDF / tables
pymupdf>=1.24.0
camelot-py>=0.11.0

# API (LangGraph service in Docker)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Shared
python-dotenv>=1.0.0
typing_extensions>=4.9.0
